+++ NEWS.md
@@ -1,3 +1,7 @@
+# rio 0.5.22
+
+* Added an `export_list()` function to write a list of data frames to multiple files using a vector of file names or a file pattern. (#207, h/t Bill Denney)
+
 # rio 0.5.21
 
 * Added support for Apache Arrow (Parquet) files. (#214)